feat: add config info and redesign layout
